remove multitenant/opensource split
