--- lib/pedant/multitenant.rb
@@ -1,2 +1,0 @@
-require 'pedant/multitenant/platform'
-require 'pedant/multitenant/rspec_config'
