+++ lib/pedant/requestor.rb
@@ -119,6 +119,7 @@ module Pedant
     end
 
     def delete!
+      puts "**** Deleting User #{self.name}"
       platform.delete_user(self)
     end
   end
@@ -131,6 +132,7 @@ module Pedant
     end
 
     def delete!
+      puts "**** Deleting Client #{self.name}"
       platform.delete_client(self)
     end
   end
