+++ lib/pedant/rspec/common.rb
@@ -493,9 +493,7 @@ module Pedant
         shared(:admin_user)       { platform.admin_user }
         shared(:normal_user)      { platform.non_admin_user }
 
-        # OSC FIXME: platform.bad_user is breaking OSC tests for some reason
-        # This is set to bad_client for now
-        shared(:outside_user)     { platform.bad_client }
+        shared(:outside_user)     { platform.bad_user}
 
         shared(:admin_client)     { platform.admin_client }
         shared(:normal_client)    { platform.non_admin_client }
