+++ lib/pedant/rspec_config.rb
@@ -1,21 +1,20 @@
-require 'pedant/multitenant/response_bodies'
-require 'pedant/multitenant/org_info'
+require 'pedant/response_bodies'
+require 'pedant/org_info'
 require 'fileutils'
 
-# Adds RSpec configuration powers to the MultiTenantPlatform
+# Adds RSpec configuration powers to the Platform
 module Pedant
-  class MultiTenantPlatform
+  class Platform
     def configure_rspec
       # Create a path for all of our generated content. Makes it easier
       # to do a final cleanup pass
       FileUtils.mkpath File.join(Dir.tmpdir, "oc-chef-pedant")
       ::RSpec.configure do |c|
-        puts "****** Configuring RSpec for Multi-Tenant Tests"
         c.run_all_when_everything_filtered = true
         c.filter_run_excluding :platform => :open_source, :intermittent_failure => true
 
-        c.include Pedant::MultiTenant::ResponseBodies
-        c.include Pedant::MultiTenant::OrgInfo
+        c.include Pedant::ResponseBodies
+        c.include Pedant::OrgInfo
       end
     end
   end
